Website redirects if there is or isnt a token cookie.
