+++ components/Loading.js
@@ -1,11 +1,12 @@
 import { useStateContext } from '@/app/stateContext';
+import styles from '@/styles/components/Loading.module.css';
 
 const Loading = () => {
   const { sidebarOpen } = useStateContext();
 
   return (
     <div
-      className={`container ${sidebarOpen ? 'container--sidebar-open' : 'container--sidebar-closed'}`}
+      className={`container ${sidebarOpen ? 'container--sidebar-open' : 'container--sidebar-closed'} ${styles.loading}`}
     ></div>
   );
 };
