+++ styles/components/Loading.module.css
@@ -1,0 +1,6 @@
+.loading {
+  position: absolute;
+  height: 100vh;
+  width: 100vw;
+  z-index: 99999999;
+}
